refactor(i18n): 将labelMap替换为动态获取标签的函数
重构国际化标签的获取方式，从静态对象改为动态函数，以支持语言切换时实时更新标签内容
