+++ src/renderer/src/components/OAuth/OAuthButton.tsx
@@ -1,4 +1,4 @@
-import { providerLabelMap } from '@renderer/i18n/labelMap'
+import { getProviderLabel } from '@renderer/i18n/label'
 import { Provider } from '@renderer/types'
 import { oauthWithAihubmix, oauthWithPPIO, oauthWithSiliconFlow, oauthWithTokenFlux } from '@renderer/utils/oauth'
 import { Button, ButtonProps } from 'antd'
@@ -40,7 +40,7 @@ const OAuthButton: FC<Props> = ({ provider, onSuccess, ...buttonProps }) => {
 
   return (
     <Button type="primary" onClick={onAuth} shape="round" {...buttonProps}>
-      {t('settings.provider.oauth.button', { provider: providerLabelMap[provider.id] })}
+      {t('settings.provider.oauth.button', { provider: getProviderLabel(provider.id) })}
     </Button>
   )
 }
