+++ src/renderer/src/components/Popups/BackupPopup.tsx
@@ -1,5 +1,5 @@
 import { loggerService } from '@logger'
-import { progressLabelMap } from '@renderer/i18n/labelMap'
+import { getProgressLabel } from '@renderer/i18n/label'
 import { backup } from '@renderer/services/BackupService'
 import store from '@renderer/store'
 import { IpcChannel } from '@shared/IpcChannel'
@@ -59,7 +59,7 @@ const PopupContainer: React.FC<Props> = ({ resolve }) => {
         progress: Math.floor(progressData.progress)
       })
     }
-    return progressLabelMap[progressData.stage]
+    return getProgressLabel(progressData.stage)
   }
 
   BackupPopup.hide = onCancel
