+++ src/renderer/src/components/Tab/TabContainer.tsx
@@ -2,7 +2,7 @@ import { PlusOutlined } from '@ant-design/icons'
 import { isLinux, isMac, isWin } from '@renderer/config/constant'
 import { useTheme } from '@renderer/context/ThemeProvider'
 import { useFullscreen } from '@renderer/hooks/useFullscreen'
-import { titleLabelMap } from '@renderer/i18n/labelMap'
+import { getTitleLabel } from '@renderer/i18n/label'
 import tabsService from '@renderer/services/TabsService'
 import { useAppDispatch, useAppSelector } from '@renderer/store'
 import type { Tab } from '@renderer/store/tabs'
@@ -24,7 +24,6 @@ import {
   X
 } from 'lucide-react'
 import { useCallback, useEffect } from 'react'
-import { useTranslation } from 'react-i18next'
 import { useLocation, useNavigate } from 'react-router-dom'
 import styled from 'styled-components'
 
@@ -63,7 +62,6 @@ let lastSettingsPath = '/settings/provider'
 const specialTabs = ['launchpad', 'settings']
 
 const TabsContainer: React.FC<TabsContainerProps> = ({ children }) => {
-  const { t } = useTranslation()
   const location = useLocation()
   const navigate = useNavigate()
   const dispatch = useAppDispatch()
@@ -135,7 +133,7 @@ const TabsContainer: React.FC<TabsContainerProps> = ({ children }) => {
               <Tab key={tab.id} active={tab.id === activeTabId} onClick={() => navigate(tab.path)}>
                 <TabHeader>
                   {tab.id && <TabIcon>{getTabIcon(tab.id)}</TabIcon>}
-                  <TabTitle>{t(titleLabelMap[tab.id])}</TabTitle>
+                  <TabTitle>{getTitleLabel(tab.id)}</TabTitle>
                 </TabHeader>
                 {tab.id !== 'home' && (
                   <CloseButton
