+++ src/renderer/src/pages/settings/ProviderSettings/index.tsx
@@ -4,7 +4,7 @@ import { loggerService } from '@logger'
 import Scrollbar from '@renderer/components/Scrollbar'
 import { getProviderLogo } from '@renderer/config/providers'
 import { useAllProviders, useProviders } from '@renderer/hooks/useProvider'
-import { providerLabelMap } from '@renderer/i18n/labelMap'
+import { getProviderLabel } from '@renderer/i18n/label'
 import ImageStorage from '@renderer/services/ImageStorage'
 import { INITIAL_PROVIDERS } from '@renderer/store/llm'
 import { Provider, ProviderType } from '@renderer/types'
@@ -102,7 +102,7 @@ const ProvidersList: FC = () => {
       }
 
       const providerDisplayName = existingProvider.isSystem
-        ? providerLabelMap[existingProvider.id]
+        ? getProviderLabel(existingProvider.id)
         : existingProvider.name
 
       // 检查是否已有 API Key
