+++ src/renderer/src/utils/__tests__/match.test.ts
@@ -83,7 +83,7 @@ describe('match', () => {
     })
 
     it('should match i18n name for system provider', () => {
-      expect(matchKeywordsInProvider('i18n:provider.sys', sysProvider)).toBe(true)
+      expect(matchKeywordsInProvider('sys', sysProvider)).toBe(true)
       expect(matchKeywordsInProvider('SystemProvider', sysProvider)).toBe(false)
     })
   })
@@ -108,8 +108,8 @@ describe('match', () => {
     })
 
     it('should match model name and i18n provider name for system provider', () => {
-      expect(matchKeywordsInModel('gpt-4.1 i18n:provider.sys', model, sysProvider)).toBe(true)
-      expect(matchKeywordsInModel('i18n:provider.sys', model, sysProvider)).toBe(true)
+      expect(matchKeywordsInModel('gpt-4.1 sys', model, sysProvider)).toBe(true)
+      expect(matchKeywordsInModel('sys', model, sysProvider)).toBe(true)
       expect(matchKeywordsInModel('SystemProvider', model, sysProvider)).toBe(false)
     })
